adapt to cpymadtools extracted API
