+++ pyhdtoolkit/cpymadtools/constants.py
@@ -21,6 +21,7 @@ MONITOR_TWISS_COLUMNS: List[str] = ["name", "s", "betx", "bety", "alfx", "alfy",
 # MQX + maybe F (1/3 in HLLHC) + A (1/3) or B (2) + . + maybe A or B (2) + triplet number (1/2/3) + side (R/L) + IP number (1/2/5/8)
 LHC_TRIPLETS_REGEX = "^MQXF?[AB].[AB]?[123][RL][1258]"
 
+# This might not be accurate anymore
 LHC_CROSSING_SCHEMES: Dict[str, Dict[str, float]] = {
     "flat": {},
     "lhc_inj": {
@@ -154,34 +155,34 @@ HLLHC_CORRECTOR_LIMITS: Dict[str, float] = {  # All values are defined as multip
     "MQSX1": 0.600 / 0.050,  # 0.6 T.m @ 50 mm in IR1&IR5
     "MQSX2": 1.360 / 0.017,  # 1.36 T @ 17 mm in IR2&IR8
     # ------------- #
-    "MCSX1": 0.050 * 2 / (0.050**2),  # 0.050 Tm @ 50 mm in IR1&IR5
-    "MCSX2": 0.028 * 2 / (0.017**2),  # 0.028 T @ 17 mm in IR2&IR8
+    "MCSX1": 0.050 * 2 / (0.050 ** 2),  # 0.050 Tm @ 50 mm in IR1&IR5
+    "MCSX2": 0.028 * 2 / (0.017 ** 2),  # 0.028 T @ 17 mm in IR2&IR8
     # ------------- #
-    "MCSSX1": 0.050 * 2 / (0.050**2),  # 0.050 Tm @ 50 mm in IR1&IR5
-    "MCSSX2": 0.11 * 2 / (0.017**2),  # 0.11 T @ 17 mm in IR2&IR8
+    "MCSSX1": 0.050 * 2 / (0.050 ** 2),  # 0.050 Tm @ 50 mm in IR1&IR5
+    "MCSSX2": 0.11 * 2 / (0.017 ** 2),  # 0.11 T @ 17 mm in IR2&IR8
     # ------------- #
-    "MCOX1": 0.030 * 6 / (0.050**3),  # 0.030 Tm @ 50 mm in IR1&IR5
-    "MCOX2": 0.045 * 6 / (0.017**3),  # 0.045 T @ 17 mm in IR2&IR8
+    "MCOX1": 0.030 * 6 / (0.050 ** 3),  # 0.030 Tm @ 50 mm in IR1&IR5
+    "MCOX2": 0.045 * 6 / (0.017 ** 3),  # 0.045 T @ 17 mm in IR2&IR8
     # ------------- #
-    "MCOSX1": 0.030 * 6 / (0.050**3),  # 0.030 Tm @ 50 mm in IR1&IR5
-    "MCOSX2": 0.048 * 6 / (0.017**3),  # 0.048 T @ 17 mm in IR2&IR8
+    "MCOSX1": 0.030 * 6 / (0.050 ** 3),  # 0.030 Tm @ 50 mm in IR1&IR5
+    "MCOSX2": 0.048 * 6 / (0.017 ** 3),  # 0.048 T @ 17 mm in IR2&IR8
     # ------------- #
-    "MCDX1": 0.030 * 24 / (0.050**4),  # 0.030 Tm @ 50 mm in IR1&IR5
+    "MCDX1": 0.030 * 24 / (0.050 ** 4),  # 0.030 Tm @ 50 mm in IR1&IR5
     # ------------- #
-    "MCDSX1": 0.030 * 24 / (0.050**4),  # 0.030 Tm @ 50 mm in IR1&IR5
+    "MCDSX1": 0.030 * 24 / (0.050 ** 4),  # 0.030 Tm @ 50 mm in IR1&IR5
     # ------------- #
-    "MCTX1": 0.07 * 120 / (0.050**5),  # 0.070 Tm @ 50 mm in IR1&IR5
-    "MCTX2": 0.01 * 120 / (0.017**5),  # 0.010 Tm @ 17 mm in IR1&IR5
+    "MCTX1": 0.07 * 120 / (0.050 ** 5),  # 0.070 Tm @ 50 mm in IR1&IR5
+    "MCTX2": 0.01 * 120 / (0.017 ** 5),  # 0.010 Tm @ 17 mm in IR1&IR5
     # ------------- #
-    "MCTSX1": 0.07 * 120 / (0.050**5),  # 0.070 Tm @ 50 mm in IR1&IR5
+    "MCTSX1": 0.07 * 120 / (0.050 ** 5),  # 0.070 Tm @ 50 mm in IR1&IR5
     "MQT": 120,  # 120 T/m
     "MQS": 120,  # 120 T/m
-    "MS": 1.280 * 2 / (0.017**2),  # 1.28 T @ 17 mm
-    "MSS": 1.280 * 2 / (0.017**2),  # 1.28 T @ 17 mm
-    "MCS": 0.471 * 2 / (0.017**2),  # 0.471 T @ 17 mm
-    "MCO": 0.040 * 6 / (0.017**3),  # 0.04 T @ 17 mm
-    "MCD": 0.100 * 24 / (0.017**4),  # 0.1 T @ 17 mm
-    "MO": 0.29 * 6 / (0.017**3),  # 0.29 T @ 17 mm
+    "MS": 1.280 * 2 / (0.017 ** 2),  # 1.28 T @ 17 mm
+    "MSS": 1.280 * 2 / (0.017 ** 2),  # 1.28 T @ 17 mm
+    "MCS": 0.471 * 2 / (0.017 ** 2),  # 0.471 T @ 17 mm
+    "MCO": 0.040 * 6 / (0.017 ** 3),  # 0.04 T @ 17 mm
+    "MCD": 0.100 * 24 / (0.017 ** 4),  # 0.1 T @ 17 mm
+    "MO": 0.29 * 6 / (0.017 ** 3),  # 0.29 T @ 17 mm
 }
 
 FD_FAMILIES: Set[str] = {"MO", "MS", "MQT"}  # Magnets that have F and D families
